refactor(common): move Environment abstraction to common module
- Move Environment interface and SystemEnvironment to jkube-kit-common
  for reuse across modules (build/api, docker service, etc.)
- Add getEnvMap() method to support bulk environment variable access
- Use initialization-on-demand holder pattern for lazy singleton
- Replace System.getenv() calls in JKubeFileInterpolator,
  KubernetesConfigAuthUtil, and OpenShiftRegistryAuthHandler
- Rename EnvironmentVariablesTestUtil to TestEnvironment and move
  to common module test-jar for sharing across modules
