+++ .gitignore
@@ -20,3 +20,4 @@ velocity.log
 /.recommenders/
 .factorypath
 bin/
+.DS_Store
